third_party/WebKit/LayoutTests/http/tests/media/media-source/mediasource-duration.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="mediasource-util.js"></script>
        <link rel='stylesheet' href='/w3c/resources/testharness.css'>
    </head>
    <body>
        <div id="log"></div>
        <script>
          function mediasource_duration_below_currentTime_seek_test(testFunction, description, options)
          {
              return mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
              {
                  assert_greater_than(segmentInfo.duration, 2, 'Sufficient test media duration');

                  var fullDuration = segmentInfo.duration;
                  var seekTo = fullDuration / 2.0;
                  var reducedDuration = seekTo / 2.0;

                  mediaElement.play();

                  // Append all the segments
                  test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer append completed');
                  test.expectEvent(mediaElement, 'playing', 'Playing triggered');
                  sourceBuffer.appendBuffer(mediaData);

                  test.waitForExpectedEvents(function()
                  {
                      assert_equals(mediaElement.duration, fullDuration, 'mediaElement duration matches fullDuration');
                      assert_equals(mediaSource.duration, fullDuration, 'mediaSource duration matches fullDuration');

                      test.expectEvent(mediaElement, 'seeking', 'seeking to seekTo');
                      test.expectEvent(mediaElement, 'timeupdate', 'timeupdate while seeking to seekTo');
                      test.expectEvent(mediaElement, 'seeked', 'seeked to seekTo');
                      mediaElement.currentTime = seekTo;
                      assert_true(mediaElement.seeking, 'mediaElement.seeking (to seekTo)');
                  });

                  test.waitForExpectedEvents(function()
                  {
                      assert_greater_than_equal(mediaElement.currentTime, seekTo, 'Playback time has reached seekTo');
                      assert_equals(mediaElement.duration, fullDuration, 'mediaElement duration matches fullDuration after seekTo');
                      assert_equals(mediaSource.duration, fullDuration, 'mediaSource duration matches fullDuration after seekTo');
                      assert_false(mediaElement.seeking, 'mediaElement.seeking after seeked to seekTo');

                      // Explicitly remove buffered media beyond the new reduced duration prior to reducing duration.
                      // Implicit removal of buffered media as part of duration reduction is disallowed as of
                      // https://github.com/w3c/media-source/pull/65/
                      test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer range removal completed');
                      sourceBuffer.remove(reducedDuration, fullDuration);
                      assert_true(sourceBuffer.updating, 'sourceBuffer.updating during range removal');
                  });

                  test.waitForExpectedEvents(function()
                  {
                      assert_false(sourceBuffer.updating, 'sourceBuffer.updating after range removal');
                      assert_greater_than_equal(mediaElement.currentTime, seekTo,
                                                'Playback time is still at least seekTo after range removal');

                      test.expectEvent(mediaElement, 'seeking', 'Seeking to reduced duration');
                      mediaSource.duration = reducedDuration;
                      assert_true(mediaElement.seeking, 'Seeking after setting reducedDuration');
                  });

                  test.waitForExpectedEvents(function()
                  {
                      assert_equals(mediaElement.currentTime, reducedDuration,
                                    'Playback time is reducedDuration while seeking');
                      assert_true(mediaElement.seeking, 'mediaElement.seeking while seeking to reducedDuration');
                      assert_equals(mediaElement.duration, reducedDuration,
                                    'mediaElement duration matches reducedDuration during seek to it');
                      assert_equals(mediaSource.duration, reducedDuration,
                                    'mediaSource duration matches reducedDuration during seek to it');

                      // FIXME: Confirm 'waiting' and then 'stalled' fire here. See http://crbug.com/266592.

                      testFunction(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData,
                                   reducedDuration);
                  });
              }, description, options);
          }

          mediasource_duration_below_currentTime_seek_test(function(test, mediaElement, mediaSource, segmentInfo,
                                                           sourceBuffer, mediaData, reducedDuration)
          {
              // Tests that duration reduction below current playback position
              // starts seek to new duration.
              test.done();
          }, 'Test seek starts on duration reduction below currentTime');

          mediasource_duration_below_currentTime_seek_test(function(test, mediaElement, mediaSource, segmentInfo,
                                                           sourceBuffer, mediaData, reducedDuration)
          {
              // The duration has been reduced at this point, and there is an
              // outstanding seek pending.
              test.expectEvent(sourceBuffer, 'updateend', 'updateend after appending more data');

              // FIXME: Confirm 'playing' fires here. See http://crbug.com/266592.

              test.expectEvent(mediaElement, 'timeupdate', 'timeupdate while finishing seek to reducedDuration');
              test.expectEvent(mediaElement, 'seeked', 'seeked to reducedDuration');

              // Allow seek to complete by appending more data beginning at the
              // reduced duration timestamp.
              sourceBuffer.timestampOffset = reducedDuration;
              sourceBuffer.appendBuffer(mediaData);

              test.waitForExpectedEvents(function()
              {
                  assert_greater_than_equal(mediaElement.currentTime, reducedDuration,
                                            'Playback time has reached reducedDuration');
                  assert_approx_equals(mediaElement.duration, reducedDuration + segmentInfo.duration, 0.05,
                                       'mediaElement duration increased by new append');
                  assert_equals(mediaSource.duration, mediaElement.duration,
                                'mediaSource duration increased by new append');
                  assert_false(mediaElement.seeking, 'mediaElement.seeking after seeked to reducedDuration');

                  test.done();
              });
          }, 'Test appendBuffer completes previous seek to reduced duration');

          mediasource_duration_below_currentTime_seek_test(function(test, mediaElement, mediaSource, segmentInfo,
                                                           sourceBuffer, mediaData, reducedDuration)
          {
              // The duration has been reduced at this point, and there is an
              // outstanding seek pending.
              test.expectEvent(mediaSource, 'sourceended', 'endOfStream acknowledged');

              // FIXME: Investigate if 'playing' should fire here. See http://crbug.com/266592.

              test.expectEvent(mediaElement, 'timeupdate', 'timeupdate while finishing seek to reducedDuration');
              test.expectEvent(mediaElement, 'seeked', 'seeked to reducedDuration');

              // Call endOfStream() to complete the pending seek.
              mediaSource.endOfStream();

              test.waitForExpectedEvents(function()
              {
                  assert_equals(mediaElement.currentTime, reducedDuration,
                                'Playback time has reached reducedDuration');
                  assert_equals(mediaElement.duration, reducedDuration,
                                'mediaElement duration matches reducedDuration after seek to it');
                  assert_equals(mediaSource.duration, reducedDuration,
                                'mediaSource duration matches reducedDuration after seek to it');
                  assert_false(mediaElement.seeking, 'mediaElement.seeking after seeked to reducedDuration');

                  test.done();
              });
          }, 'Test endOfStream completes previous seek to reduced duration');

          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              assert_greater_than(segmentInfo.duration, 2, 'Sufficient test media duration');

              var fullDuration = segmentInfo.duration;
              var newDuration = fullDuration * 2;

              var durationchangeEventCounter = 0;
              var expectedDurationChangeEventCount = 1;
              var durationchangeEventHandler = test.step_func(function(event)
              {
                  assert_equals(mediaElement.duration, newDuration, 'mediaElement duration matches newDuration');
                  assert_equals(mediaSource.duration, newDuration, 'mediaSource duration matches newDuration');
                  durationchangeEventCounter++;
              });

              mediaElement.play();

              // Append all the segments
              test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer append completed');
              test.expectEvent(mediaElement, 'playing', 'Playing triggered');
              sourceBuffer.appendBuffer(mediaData);

              test.waitForExpectedEvents(function()
              {
                  assert_equals(mediaElement.duration, fullDuration, 'mediaElement duration matches fullDuration');
                  assert_equals(mediaSource.duration, fullDuration, 'mediaSource duration matches fullDuration');
                  assert_less_than(mediaElement.currentTime, newDuration / 2, 'mediaElement currentTime check');

                  // Media load also fires 'durationchange' event, so only start counting them now.
                  mediaElement.addEventListener('durationchange', durationchangeEventHandler);

                  // Increase duration. This should result in one 'durationchange' fired.
                  mediaSource.duration = newDuration;
                  assert_false(sourceBuffer.updating, "sourceBuffer.updating after duration set to newDuration");

                  // Set duration again to make sure it does not trigger another 'durationchange' event.
                  mediaSource.duration = newDuration;

                  assert_false(sourceBuffer.updating, "sourceBuffer.updating after duration set again to newDuration");

                  // Mark endOfStream so that playback can reach 'ended' at the new duration.
                  test.expectEvent(mediaSource, 'sourceended', 'endOfStream acknowledged');
                  mediaSource.endOfStream();

                  // endOfStream can change duration downwards slightly.
                  // Allow for one more 'durationchange' event only in this case.
                  var currentDuration = mediaSource.duration;
                  if (currentDuration != newDuration) {
                      assert_true(currentDuration > 0 && currentDuration < newDuration, 'adjusted duration check');
                      newDuration = currentDuration;
                      ++expectedDurationChangeEventCount;
                  }
              });

              test.waitForExpectedEvents(function()
              {
                  // Allow any remaining queued durationchange to fire, while counting 'durationchange' them.
                  test.step_timeout(test.step_func(function()
                  {
                      mediaElement.removeEventListener('durationchange', durationchangeEventHandler);
                      assert_equals(durationchangeEventCounter, expectedDurationChangeEventCount, 'durationchange count check');
                      test.done();
                  }), 0);
              });
          }, 'Test setting same duration multiple times does not fire duplicate durationchange', {timeout: 2500});

          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              assert_greater_than(segmentInfo.duration, 2, 'Sufficient test media duration');

              var fullDuration = segmentInfo.duration;

              // Append all the segments
              test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer');
              sourceBuffer.appendBuffer(mediaData);

              test.waitForExpectedEvents(function()
              {
                  assert_equals(mediaElement.duration, fullDuration, 'mediaElement duration matches fullDuration');
                  assert_equals(mediaSource.duration, fullDuration, 'mediaSource duration matches fullDuration');

                  // TODO(wolenetz): Fine-tune this test to use the buffered attribute's highest end time
                  // instead of fullDuration once Chrome correctly reports buffered PTS, not DTS.
                  // See https://crbug.com/398130.

                  // Setting duration to same as current, or increasing it, should not trigger exception.
                  mediaSource.duration = fullDuration;
                  mediaSource.duration = fullDuration + 1;

                  // Reducing duration to below the highest buffered PTS should trigger exception.
                  assert_throws('InvalidStateError',
                      function() { mediaSource.duration = fullDuration - 0.05; },
                      'Duration reduction that truncates at least one whole coded frame throws an exception.');

                  assert_equals(mediaSource.duration, fullDuration + 1, 'mediaSource duration matches fullDuration+1');

                  // Reducing duration without truncating any buffered media should not trigger exception.
                  mediaSource.duration = fullDuration;

                  // Reducing duration by less than the minimum of the last test audio and video frame
                  // durations should not trigger exception.
                  mediaSource.duration = fullDuration - 0.001;
                  test.done();
              });
          }, 'Test duration reduction below highest buffered presentation time is disallowed');

        </script>
    </body>
</html>
